requirements.txt
google-generativeai
# Not sure about the above one
pandas
opencv-python
numpy
scikit-learn
tensorflow
requests
Pillow
transformers
opencv-python
numpy

# yolov4.weights
# yolov4.cfg
# coco.names
